# apps/server/openapi.yaml
openapi: 3.1.0
paths:
  /api/auth/sign-up:
    post:
      operationId: signUp
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignUpDTO"
      responses:
        "201":
          description: ""
      summary: ""
      tags: &a1
        - auth
  /api/auth/login/credentials:
    post:
      operationId: credentialsLogin
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialsLoginDTO"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JwtTokenDTO"
        "401":
          description: Email or Password not exits or matches
      summary: ""
      tags: *a1
  /api/auth/session:
    get:
      operationId: session
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionDTO"
      security:
        - bearer: []
      summary: ""
      tags: *a1
  /api/auth/users:
    get:
      operationId: getAllPassengers
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublicUserDTO"
      summary: ""
      tags: *a1
  /api/auth/logout:
    post:
      operationId: logout
      parameters: []
      responses:
        "201":
          description: ""
      summary: ""
      tags: *a1
  /api/ride/estimate:
    post:
      description: Estimate Ride / Calcular corrida
      operationId: estimateRide
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EstimateRideDTO"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstimateRideCreatedDTO"
      summary: estimateRide
      tags: &a2
        - ride
  /api/ride/confirm:
    patch:
      description: Confirm Ride / Confirmar corrida
      operationId: confirmRide
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EstimateRideDTO"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstimateRideCreatedDTO"
      summary: confirmRide
      tags: *a2
  /api/ride/address:
    get:
      description: Estimate Ride / Calcular corrida
      operationId: getAddress
      parameters:
        - name: street
          required: true
          in: query
          description: Street name
          schema:
            type: string
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
      summary: getAddress
      tags: *a2
info:
  title: Shooper Travel API
  description: Server API
  version: 1.0.0
  contact: {}
tags:
  - name: auth
    description: Auth API
  - name: users
    description: Users API
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    SignUpDTO:
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
          minLength: 3
        password:
          type: string
        role:
          type: string
          enum:
            - passenger
            - driver
          x-enumNames:
            - passenger
            - driver
      required:
        - email
        - name
        - password
        - role
    CredentialsLoginDTO:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required:
        - email
        - password
    JwtTokenDTO:
      type: object
      properties:
        type:
          description: Token Type
          type: string
          enum:
            - Bearer
          x-enumNames:
            - BEARER
          default: Bearer
        token:
          description: JWT Token Type
          type: string
        expiryAt:
          description: Expiration At
          type: string
      required:
        - token
        - expiryAt
    SessionDTO:
      description: Account Session Data
      type: object
      properties:
        id:
          type: string
          format: uuid
        role:
          type: string
          enum:
            - passenger
            - driver
          x-enumNames:
            - passenger
            - driver
        email:
          type: string
          format: email
        name:
          type: string
        iss:
          type: string
        iat:
          type: number
        exp:
          type: number
      required:
        - id
        - role
        - email
        - name
        - iss
        - iat
        - exp
    PublicUserDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
          minLength: 3
      required:
        - id
        - email
        - name
    EstimateRideDTO:
      type: object
      properties:
        customer_id:
          type: string
          format: uuid
        origin:
          type: string
          minLength: 4
        destination:
          type: string
          minLength: 4
      required:
        - customer_id
        - origin
        - destination
    EstimateRideCreatedDTO:
      type: object
      properties:
        origin:
          type: object
          properties:
            latitude:
              type: number
            longitude:
              type: number
          required:
            - latitude
            - longitude
        destination:
          type: object
          properties:
            latitude:
              type: number
            longitude:
              type: number
          required:
            - latitude
            - longitude
        distance:
          type: number
        duration:
          type: string
        options:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              description:
                type: string
              vehicle:
                type: string
              review:
                type: object
                properties:
                  rating:
                    type: number
                  comment:
                    type: string
                required:
                  - rating
                  - comment
              value:
                type: number
            required:
              - id
              - name
              - description
              - vehicle
              - review
              - value
        routeResponse:
          type: object
          additionalProperties: {}
      required:
        - origin
        - destination
        - distance
        - duration
        - options
        - routeResponse